feat: implement identity access constraints, tests yet to be change

+++ docuisine/routes/category.py
@@ -1,9 +1,10 @@
 from fastapi import APIRouter, HTTPException, status
 
 from docuisine.db.models import Category
-from docuisine.dependencies import Category_Service
+from docuisine.dependencies import AuthenticatedUser, Category_Service
 from docuisine.schemas.category import CategoryCreate, CategoryOut, CategoryUpdate
 from docuisine.schemas.common import Detail
+from docuisine.schemas.enums import Role
 from docuisine.utils import errors
 
 router = APIRouter(prefix="/categories", tags=["Categories"])
@@ -11,7 +12,11 @@ router = APIRouter(prefix="/categories", tags=["Categories"])
 
 @router.get("/", status_code=status.HTTP_200_OK, response_model=list[CategoryOut])
 async def get_categories(category_service: Category_Service) -> list[CategoryOut]:
-    """Get all categories."""
+    """
+    Get all categories.
+
+    Access Level: Public
+    """
     categories: list[Category] = category_service.get_all_categories()
     return [CategoryOut.model_validate(category) for category in categories]
 
@@ -23,7 +28,11 @@ async def get_categories(category_service: Category_Service) -> list[CategoryOut
     responses={status.HTTP_404_NOT_FOUND: {"model": Detail}},
 )
 async def get_category(category_id: int, category_service: Category_Service) -> CategoryOut:
-    """Get a category by ID."""
+    """
+    Get a category by ID.
+
+    Access Level: Public
+    """
     try:
         category: Category = category_service.get_category(category_id=category_id)
         return CategoryOut.model_validate(category)
@@ -41,9 +50,17 @@ async def get_category(category_id: int, category_service: Category_Service) -> 
     responses={status.HTTP_409_CONFLICT: {"model": Detail}},
 )
 async def create_category(
-    category: CategoryCreate, category_service: Category_Service
+    category: CategoryCreate,
+    category_service: Category_Service,
+    authenticated_user: AuthenticatedUser,
 ) -> CategoryOut:
-    """Create a new category."""
+    """
+    Create a new category.
+
+    Access Level: Admin
+    """
+    if authenticated_user.role != Role.ADMIN:
+        raise errors.ForbiddenAccessError
     try:
         new_category: Category = category_service.create_category(
             name=category.name, description=category.description
@@ -61,14 +78,24 @@ async def create_category(
     status_code=status.HTTP_200_OK,
     response_model=CategoryOut,
     responses={
+        status.HTTP_403_FORBIDDEN: {"model": Detail},
         status.HTTP_404_NOT_FOUND: {"model": Detail},
         status.HTTP_409_CONFLICT: {"model": Detail},
     },
 )
 async def update_category(
-    category_id: int, category: CategoryUpdate, category_service: Category_Service
+    category_id: int,
+    category: CategoryUpdate,
+    category_service: Category_Service,
+    authenticated_user: AuthenticatedUser,
 ) -> CategoryOut:
-    """Update a category by ID."""
+    """
+    Update an existing category.
+
+    Access Level: Admin
+    """
+    if authenticated_user.role != Role.ADMIN:
+        raise errors.ForbiddenAccessError
     try:
         updated_category: Category = category_service.update_category(
             category_id=category_id, name=category.name, description=category.description
@@ -92,8 +119,16 @@ async def update_category(
     response_model=Detail,
     responses={status.HTTP_404_NOT_FOUND: {"model": Detail}},
 )
-async def delete_category(category_id: int, category_service: Category_Service) -> Detail:
-    """Delete a category by ID."""
+async def delete_category(
+    category_id: int, category_service: Category_Service, authenticated_user: AuthenticatedUser
+) -> Detail:
+    """
+    Delete a category by ID.
+
+    Access Level: Admin
+    """
+    if authenticated_user.role != Role.ADMIN:
+        raise errors.ForbiddenAccessError
     try:
         category_service.delete_category(category_id=category_id)
         return Detail(detail=f"Category with ID {category_id} has been deleted.")
