+++ docuisine/routes/recipe.py
@@ -1,8 +1,9 @@
 from fastapi import APIRouter, HTTPException, status
 
 from docuisine.db.models import Recipe
-from docuisine.dependencies import Recipe_Service
+from docuisine.dependencies import AuthenticatedUser, Recipe_Service
 from docuisine.schemas.common import Detail
+from docuisine.schemas.enums import Role
 from docuisine.schemas.recipe import RecipeCreate, RecipeOut, RecipeUpdate
 from docuisine.utils import errors
 
@@ -11,7 +12,11 @@ router = APIRouter(prefix="/recipes", tags=["Recipes"])
 
 @router.get("/", status_code=status.HTTP_200_OK, response_model=list[RecipeOut])
 async def get_recipes(recipe_service: Recipe_Service) -> list[RecipeOut]:
-    """Get all recipes."""
+    """
+    Get all recipes.
+
+    Access Level: Public
+    """
     recipes: list[Recipe] = recipe_service.get_all_recipes()
     return [RecipeOut.model_validate(recipe) for recipe in recipes]
 
@@ -22,7 +27,11 @@ async def get_recipes(recipe_service: Recipe_Service) -> list[RecipeOut]:
     response_model=list[RecipeOut],
 )
 async def get_recipes_by_user(user_id: int, recipe_service: Recipe_Service) -> list[RecipeOut]:
-    """Get all recipes created by a specific user."""
+    """
+    Get all recipes created by a specific user.
+
+    Access Level: Public
+    """
     recipes: list[Recipe] = recipe_service.get_recipes_by_user(user_id=user_id)
     return [RecipeOut.model_validate(recipe) for recipe in recipes]
 
@@ -34,7 +43,11 @@ async def get_recipes_by_user(user_id: int, recipe_service: Recipe_Service) -> l
     responses={status.HTTP_404_NOT_FOUND: {"model": Detail}},
 )
 async def get_recipe(recipe_id: int, recipe_service: Recipe_Service) -> RecipeOut:
-    """Get a recipe by ID."""
+    """
+    Get a recipe by ID.
+
+    Access Level: Public
+    """
     try:
         recipe: Recipe = recipe_service.get_recipe(recipe_id=recipe_id)
         return RecipeOut.model_validate(recipe)
@@ -48,8 +61,18 @@ async def get_recipe(recipe_id: int, recipe_service: Recipe_Service) -> RecipeOu
     response_model=RecipeOut,
     responses={status.HTTP_409_CONFLICT: {"model": Detail}},
 )
-async def create_recipe(recipe: RecipeCreate, recipe_service: Recipe_Service) -> RecipeOut:
-    """Create a new recipe."""
+async def create_recipe(
+    recipe: RecipeCreate,
+    recipe_service: Recipe_Service,
+    authenticated_user: AuthenticatedUser,
+) -> RecipeOut:
+    """
+    Create a new recipe.
+
+    Access Level: Admin, User
+    """
+    if authenticated_user.role not in {Role.ADMIN, Role.USER}:
+        raise errors.ForbiddenAccessError
     try:
         new_recipe: Recipe = recipe_service.create_recipe(
             user_id=recipe.user_id,
@@ -75,9 +98,18 @@ async def create_recipe(recipe: RecipeCreate, recipe_service: Recipe_Service) ->
     },
 )
 async def update_recipe(
-    recipe_id: int, recipe: RecipeUpdate, recipe_service: Recipe_Service
+    recipe_id: int,
+    recipe: RecipeUpdate,
+    recipe_service: Recipe_Service,
+    authenticated_user: AuthenticatedUser,
 ) -> RecipeOut:
-    """Update a recipe by ID."""
+    """
+    Update a recipe by ID.
+
+    Access Level: Admin, User
+    """
+    if authenticated_user.role not in {Role.ADMIN, Role.USER}:
+        raise errors.ForbiddenAccessError
     try:
         updated: Recipe = recipe_service.update_recipe(
             recipe_id=recipe_id,
@@ -101,8 +133,16 @@ async def update_recipe(
     response_model=Detail,
     responses={status.HTTP_404_NOT_FOUND: {"model": Detail}},
 )
-async def delete_recipe(recipe_id: int, recipe_service: Recipe_Service) -> Detail:
-    """Delete a recipe by ID."""
+async def delete_recipe(
+    recipe_id: int, recipe_service: Recipe_Service, authenticated_user: AuthenticatedUser
+) -> Detail:
+    """
+    Delete a recipe by ID.
+
+    Access Level: Admin, User
+    """
+    if authenticated_user.role not in {Role.ADMIN, Role.USER}:
+        raise errors.ForbiddenAccessError
     try:
         recipe_service.delete_recipe(recipe_id=recipe_id)
         return Detail(detail=f"Recipe with ID {recipe_id} has been deleted.")
